+++ data/duke.txt
@@ -1,0 +1,2 @@
+T | 0 | eat
+D | 0 | bath | 2029-05-30
