+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.exception.ChatException;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
 public class DeleteCommand extends Command{
     private int taskNumber;
     public DeleteCommand(int taskNumber) {
@@ -5,8 +12,8 @@ public class DeleteCommand extends Command{
     }
     public void execute(TaskList tasks, Ui ui, Storage storage){
         try {
-            tasks.deleteTask(taskNumber);
             ui.deleteTaskResponse(tasks.getTask(taskNumber), tasks);
+            tasks.deleteTask(taskNumber);
             storage.saveList(tasks);
         } catch (ChatException e) {
             ui.showLoadingError(e);
