+++ src/main/java/duke/command/MarkCommand.java
@@ -1,12 +1,19 @@
-public class UnmarkCommand extends Command {
+package duke.command;
+
+import duke.exception.ChatException;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class MarkCommand extends Command{
     int taskNumber;
-    public UnmarkCommand(int taskNumber) {
+    public MarkCommand(int taskNumber) {
         this.taskNumber = taskNumber;
     }
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
-            tasks.markUndone(taskNumber);
-            ui.markUndoneResponse(tasks.getTask(taskNumber));
+            tasks.markDone(taskNumber);
+            ui.markDoneResponse(tasks.getTask(taskNumber));
             storage.saveList(tasks);
         } catch (ChatException e) {
             ui.showLoadingError(e);
