+++ src/main/java/duke/exception/ChatException.java
@@ -1,3 +1,5 @@
+package duke.exception;
+
 public class ChatException extends Exception {
     public ChatException(String errorMessage) {
         super(errorMessage);
